[#11] Rename `neofs` contract to `frostfs`
