+++ pkg/innerring/innerring.go
@@ -13,8 +13,8 @@ import (
 	"github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/audit"
 	"github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/balance"
 	"github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/container"
+	frostfs "github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/frostfs"
 	"github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/governance"
-	frostfs "github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/neofs"
 	"github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/netmap"
 	nodevalidator "github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/netmap/nodevalidation"
 	addrvalidator "github.com/TrueCloudLab/frostfs-node/pkg/innerring/processors/netmap/nodevalidation/maddress"
@@ -29,7 +29,7 @@ import (
 	auditClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/audit"
 	balanceClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/balance"
 	cntClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/container"
-	frostfsClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/neofs"
+	frostfsClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/frostfs"
 	"github.com/TrueCloudLab/frostfs-node/pkg/morph/client/neofsid"
 	nmClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/netmap"
 	repClient "github.com/TrueCloudLab/frostfs-node/pkg/morph/client/reputation"
