+++ pkg/morph/client/frostfs/bind.go
