+++ pkg/morph/event/frostfs/bind.go
