+++ pkg/morph/event/frostfs/bind_test.go
