+++ pkg/morph/event/frostfs/config.go
