+++ pkg/morph/event/frostfs/config_test.go
