+++ pkg/morph/event/frostfs/deposit.go
