+++ pkg/morph/event/frostfs/ir_update.go
