+++ pkg/morph/event/frostfs/ir_update_test.go
