+++ pkg/morph/event/frostfs/withdraw.go
