+++ pkg/morph/event/frostfs/withdraw_test.go
